fix: round 2 int64 migration - indexing and loop variables
Fancy indexing:
- NDArray.Indexing.Selection.Getter.cs: computedOffsets NDArray<int> -> NDArray<long>
- NDArray.Indexing.Selection.Setter.cs: same change
- NDArray.Indexing.Selection.cs: GetIndicesFromSlice returns NDArray<long>
- Updated int* pointers to long*

Loop variables:
- np.array.cs, np.all.cs, np.any.cs: int -> long loop vars
- ndarray.argsort.cs: added size checks for Enumerable.Range
- NDArray.NOT.cs, NDArray.Equals.cs: int -> long loop vars

+++ src/NumSharp.Core/Creation/np.array.cs
@@ -112,7 +112,7 @@ namespace NumSharp
                     Func<bool> next = enumerator.MoveNext;
 
                     var addr = slice.Address;
-                    for (int i = 0; i < size && next(); i++)
+                    for (long i = 0; i < size && next(); i++)
                     {
                         addr[i] = enumerator.Current;
                     }
